refactor(livenessgate): move into subtreevalidation, drop redundant wrapper
The liveness gate is a feature-specific decision helper, not a general
teranode utility — both callers (subtreevalidation per-subtree gating
and blockvalidation catchup prefetch) consume the same one feature,
driven by SubtreeValidation.AssumeTxsBroadcastToAllNodes.

  - Move util/livenessgate/ → services/subtreevalidation/livenessgate/
    so the package lives alongside the setting that owns it.
  - Drop the package-level ShouldUseSubtreeOnlyPath convenience. Callers
    use Decide directly and compare against DecisionSubtreeOnly (exactly
    one extra line at the one site that used the wrapper).
  - Drop the three `var _ livenessgate.Client = (*T)(nil)` assertions in
    services/blockchain/. They created a reverse-layering import
    (blockchain depending on a subtreevalidation subpackage) and were
    redundant — real call sites already pass the blockchain client to
    livenessgate.Decide, so the compiler checks the interface match at
    the call site.

Behaviour unchanged. The Decision enum + metric labels are kept; the
(*Server).ShouldUseSubtreeOnlyPath wrapper in subtreevalidation is
kept (it bundles metric emission and logging over Decide).

Net: -37 lines, one fewer package, no cross-service dependency inversion.
